PCHR-513: Draft of custom bootstrap-base admin theme
The theme is a subtheme of the Bootstrap theme (https://www.drupal.org/project/bootstrap)

While the Bootstrap theme uses the Bootstrap's cdn directly, the subtheme relies on the entire Bootstrap source to be present in the bootstrap/ folder. It then has a less/ folder which contains some duplicates (not sure why) and then a style.less file that bundle everything together

I've added less and gulp to the theme. By running `gulp` or `gulp watch`, any changes to the files in less/ trigger the re-building of css/style.css